Split trio.testing into submodules
Should rearrange the tests a bit too, but this gets rid of the
horrible 1438 line file.

Also: run yapf.
