+++ trio/testing/_trio_test.py
@@ -1,0 +1,30 @@
+from functools import wraps, partial
+
+from .. import _core
+from ..abc import Clock
+
+__all__ = ["trio_test"]
+
+
+# Use:
+#
+#    @trio_test
+#    async def test_whatever():
+#        await ...
+#
+# Also: if a pytest fixture is passed in that subclasses the Clock abc, then
+# that clock is passed to trio.run().
+def trio_test(fn):
+    @wraps(fn)
+    def wrapper(**kwargs):
+        __tracebackhide__ = True
+        clocks = [c for c in kwargs.values() if isinstance(c, Clock)]
+        if not clocks:
+            clock = None
+        elif len(clocks) == 1:
+            clock = clocks[0]
+        else:
+            raise ValueError("too many clocks spoil the broth!")
+        return _core.run(partial(fn, **kwargs), clock=clock)
+
+    return wrapper
